Move main screen to a saperated package
