+++ app/src/main/AndroidManifest.xml
@@ -30,7 +30,7 @@
             android:name=".CategoriesEditorActivity"
             android:screenOrientation="portrait" />
         <activity
-            android:name=".MainActivity"
+            android:name=".screens.main.MainActivity"
             android:launchMode="singleTask"
             android:screenOrientation="portrait"
             android:theme="@style/AppTheme.NoActionBar">
