+++ app/src/main/java/com/yurii/youtubemusic/MediaItemsFragment.kt
@@ -21,7 +21,7 @@ import com.yurii.youtubemusic.ui.SelectCategoriesDialog
 import com.yurii.youtubemusic.utilities.Injector
 import com.yurii.youtubemusic.adapters.MediaListAdapter
 import com.yurii.youtubemusic.adapters.MediaListAdapterController
-import com.yurii.youtubemusic.viewmodels.MainActivityViewModel
+import com.yurii.youtubemusic.screens.main.MainActivityViewModel
 import com.yurii.youtubemusic.viewmodels.MediaItemsViewModel
 
 class MediaItemsFragment : Fragment() {
